[Kernel]: USB MSC handles SCSIInquiry command
USB MSC driver can successfully deliver and read SCSIInquiry command

Thankyou,
Manas Kamal Choudhury, 01-04-2025 (11:53 A.M)

+++ Drivers/xHCI/xhci.h
@@ -225,7 +225,7 @@ typedef struct _endp_ {
 	uint8_t endpoint_num;
 	uint8_t endpoint_type;
 	uint8_t interval;
-	uint8_t max_packet_sz;
+	uint16_t max_packet_sz;
 	unsigned cmd_ring_index;
 	unsigned cmd_ring_max;
 	unsigned cmd_ring_cycle;
